requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
pyyaml>=6.0
shapely>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
pymupdf>=1.23.0
pdf2image>=1.17.0
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.30.0
